// src/app/api/generated/models/products-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductsDto {
  categoryID?: (number | string | null);
  discontinued?: boolean;
  productID: (number | string);
  productName: (string | null);
  quantityPerUnit?: (string | null);
  reorderLevel?: (number | string);
  supplierID?: (number | string | null);
  unitPrice?: (number | string | null);
  unitsInStock?: (number | string);
  unitsOnOrder?: (number | string);
}